requirements.txt
# Core chess libraries
python-chess>=1.10.0
chess-openings>=0.1.0

# Machine learning and data processing
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data handling
pyyaml>=6.0
tqdm>=4.65.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: RL libraries for future phases
# gym>=0.26.0
# stable-baselines3>=2.0.0 